// temper/game-items-alchemy/src/generated/temper-potion-crafted.generated.ts
export const POTIONS_CRAFTED = {
  "protection-resistance-physical-resistance-spell": {
    id: "protection-resistance-physical-resistance-spell" as const,
    name: "Essence of Physical Resistance",
    description: "Grants Increase Physical Resistance, Protection, Increase Spell Resistance.",
    categoryId: "potions" as const,
    subcategoryId: "crafted" as const,
    itemId: 0,
    icon: "/esoui/art/icons/consumable_potion_007_type_005.dds",
    level: "CP150",
    seconds: 47,
    reagents: [["Beetle Scuttle", "Bugloss", "Mudcrab Chitin"], ["Beetle Scuttle", "Clam Gall", "Mudcrab Chitin"], ["Beetle Scuttle", "Mudcrab Chitin", "White Cap"]],
    effects: [
      {"metricId":"resistance-physical","effectType":"number-for-seconds","effectValue":{"value":5280,"seconds":36.6}},
      {"metricId":"resistance-spell","effectType":"number-for-seconds","effectValue":{"value":5280,"seconds":36.6}},
      {"buffId":"minor-protection","seconds":12},
    ],
  },
// ...
} satisfies Record<string, PotionsTemplate>